Server/requirements.txt
flask
flask_cors
bcrypt
pymongo
Flask-Session
python-dotenv
gunicorn
regex
